Renamed not a time to semantic time.
