+++ tests/semantic_time.py
@@ -1,38 +1,38 @@
 #!/usr/bin/python
 # -*- coding: utf-8 -*-
-"""Tests for the not a time implementation."""
+"""Tests for the semantic time implementation."""
 
 import unittest
 
-from dfdatetime import not_a_time
+from dfdatetime import semantic_time
 
 
-class NotATimeTest(unittest.TestCase):
-  """Tests for not a time."""
+class SemanticTimeTest(unittest.TestCase):
+  """Tests for semantic time."""
 
   # pylint: disable=protected-access
 
   def testCopyFromString(self):
     """Tests the CopyFromString function."""
-    not_a_time_object = not_a_time.NotATime()
+    semantic_time_object = semantic_time.SemanticTime()
 
-    with self.assertRaises(ValueError):
-      not_a_time_object.CopyFromString(u'2010-08-12 21:06:31.546875+01:00')
+    semantic_time_object.CopyFromString(u'Never')
+    self.assertEqual(semantic_time_object.string, u'Never')
 
   def testCopyToStatTimeTuple(self):
     """Tests the CopyToStatTimeTuple function."""
-    not_a_time_object = not_a_time.NotATime()
+    semantic_time_object = semantic_time.SemanticTime()
 
     expected_stat_time_tuple = (0, 0)
-    stat_time_tuple = not_a_time_object.CopyToStatTimeTuple()
+    stat_time_tuple = semantic_time_object.CopyToStatTimeTuple()
     self.assertEqual(stat_time_tuple, expected_stat_time_tuple)
 
   def testGetPlasoTimestamp(self):
     """Tests the GetPlasoTimestamp function."""
-    not_a_time_object = not_a_time.NotATime()
+    semantic_time_object = semantic_time.SemanticTime()
 
     expected_micro_posix_timestamp = 0
-    micro_posix_timestamp = not_a_time_object.GetPlasoTimestamp()
+    micro_posix_timestamp = semantic_time_object.GetPlasoTimestamp()
     self.assertEqual(micro_posix_timestamp, expected_micro_posix_timestamp)
